LayoutTests/media/media-play-promise-resolve-expected.txt


RUN(mediaElement.src = findMediaFile("video", "content/test"))
RUN(mediaElement.play().then(promiseResolved).catch(failTest))
Promise resolved. OK
END OF TEST
